Moving InsightsWrapper to own place
