+++ frontend/src/scenes/web-analytics/EventsHeatMap/EventsHeatMap.tsx
@@ -1,10 +1,10 @@
 import './EventsHeatMap.scss'
 
 import { useValues } from 'kea'
-import { InsightsWrapper } from 'products/revenue_analytics/frontend/nodes/utils'
 import React, { useCallback, useEffect, useState } from 'react'
 import { dataThemeLogic } from 'scenes/dataThemeLogic'
 import { InsightLoadingState } from 'scenes/insights/EmptyStates'
+import { InsightsWrapper } from 'scenes/insights/InsightsWrapper/InsightsWrapper'
 
 import { useResizeObserver } from '~/lib/hooks/useResizeObserver'
 import { dataNodeLogic } from '~/queries/nodes/DataNode/dataNodeLogic'
