+++ products/revenue_analytics/frontend/nodes/RevenueAnalyticsGrowthRateNode.tsx
@@ -2,6 +2,7 @@ import { BindLogic, useValues } from 'kea'
 import { useState } from 'react'
 import { InsightLoadingState } from 'scenes/insights/EmptyStates'
 import { insightLogic } from 'scenes/insights/insightLogic'
+import { InsightsWrapper } from 'scenes/insights/InsightsWrapper/InsightsWrapper'
 import { insightVizDataLogic } from 'scenes/insights/insightVizDataLogic'
 import { LineGraph } from 'scenes/insights/views/LineGraph/LineGraph'
 
@@ -13,8 +14,6 @@ import {
 } from '~/queries/schema/schema-general'
 import { QueryContext } from '~/queries/types'
 import { GraphDataset, GraphType } from '~/types'
-
-import { InsightsWrapper } from './utils'
 
 let uniqueNode = 0
 export function RevenueAnalyticsGrowthRateNode(props: {
